demos/hello-layers/requirements.txt
docker>=4.0.0
PyYAML>=5.3
